Change: method for constructing networks
Major refactor for how to construct networks. Moving away from direct
initialization to using `from_dir` and `from_data` constructors. With the
related changes to the Node and Edge classes, this should make it easier to
modify the network (especially for adding in new nodes and edges later, such as
those that are generated).

Breaks up parts of the initialization process into smaller functions which
should make things more flexible and allows us to reuse parts of the initial
construction for adding and dropping nodes and edges later.

Should hopefully also make it easier to implement extra features such as edges
that are bipartite and don't fit the expected naming scheme as well since the
logic for edges is better contained in small functions.