Change public API to be more consistent with graph names
- `from_dir` and `to_dir` functions were not consistent since `from_dir`
commands had to specify a root node and `to_dir` do not.
- So move root node to a keyword parameter so it's not the first argument to
`from_dir` (this wasn't clear anyway since it wasn't intuitive what the unnamed
argument was). (Also give root a default of publication since that is a good
default for this networks, using another root is an exception.)
- Also shift graph name to front of argument list so it reads as: read graph
"graph name" followed by optional modifiers.