Move poetry setup to poetry2nix
Instead of custom nix code to generate the pyproject.toml. This adds the
poetry.lock file to the git repo and ensures that nix produces an equivalent
environment to that of poetry for reprehensibility between computers regardless
of use of poetry or nix. This will also allow CI to use the repo's poetry.lock
instead of producing its own.