Temporarily remove check if graph with requested condition exists
Eventually `from_pubnet` should check if a graph with the given name exists and
if that graph matches the structure requested. If so it'll return that graph,
otherwise it'll error to prevent accidentally overwriting an existing graph.

Before doing this, I want to change pubmedparser to make it easier to reuse
already parsed data so I can try to update existing networks or return a subset
of a larger network.

This is low priority for now. Leaving for later.